refactor: 전역상태 type 분리
