+++ frontend/src/store/NavbarStore.ts
@@ -1,14 +1,6 @@
 import { BOOST_CAMP_OPTION, GENERATION_VALUE } from '@constant/NavbarSelect';
+import { NavbarState } from '@type/Navbar';
 import { create } from 'zustand';
-
-type NavbarState = {
-  generation: string;
-  selectedGroup: string;
-  isNavbarOpen: boolean;
-  setGeneration: (generation: string) => void;
-  setSelectedGroup: (group: string) => void;
-  toggleNavbar: (isOpen?: boolean) => void;
-};
 
 const useNavbarStore = create<NavbarState>((set) => ({
   generation: GENERATION_VALUE.NINTH,
